requirements.txt
torch
torchvision
numpy==1.12.1
scipy==0.19.0
termcolor==1.1.0
pyzmq==16.0.2
flatbuffers==2015.12.22.1
visdom==0.1.04
matplotlib==2.0.0
py-cpuinfo==3.2.0
pylatex==1.2.1
pydotplus==2.0.2
seaborn==0.7.1
